requirements.txt
PyYAML>=6.0.1,<7.0
